Lab 3: Vectorization with Intel Compilers

+++ lab3/compress.c
@@ -2,14 +2,14 @@
 // Filters positive values from array a into array b
 //
 // This pattern is recognized by the compiler and can be vectorized
-// using AVX-512's vcompresspd instruction, but not with AVX2
+// using AVX-512's vcompressps instruction, but not with AVX2
 
-int compress(double *a, double *b, int na)
+int compress(float *a, float *b, int na)
 {
     int nb = 0;
     for (int ia = 0; ia < na; ia++)
     {
-        if (a[ia] > 0.)
+        if (a[ia] > 0.f)
             b[nb++] = a[ia];
     }
